# faster `rand(1:n)` by outlining unlikely branch