--- a/stdlib/Random/src/generation.jl
+++ b/stdlib/Random/src/generation.jl
@@ -375,15 +375,14 @@ function rand(rng::AbstractRNG, sp::SamplerRangeNDL{U,T}) where {U,T}
     s = sp.s
     x = widen(rand(rng, U))
     m = x * s
-    r = (m % U) < s ? rand_unlikely(rng, s, m) % T :
-        iszero(s)   ? x % T :
-                      (m >> (8*sizeof(U))) % T
+    r::T = (m % U) < s ? rand_unlikely(rng, s, m) % T :
+           iszero(s)   ? x % T :
+                         (m >> (8*sizeof(U))) % T
     r + sp.a
 end
 
-# similar to `randn_unlikely` : splitting out this unlikely path leads to faster code
+# similar to `randn_unlikely` : splitting this unlikely path out results in faster code
 @noinline function rand_unlikely(rng, s::U, m)::U where {U}
-    # m needs to be passed, because the while loop below might have no iterations
     t = mod(-s, s) # as s is unsigned, -s is equal to 2^L - s in the paper
     while (m % U) < t
         x = widen(rand(rng, U))