--- a/test/testhelpers/coverage_file.info
+++ b/test/testhelpers/coverage_file.info
@@ -1,6 +1,6 @@
 SF:<FILENAME>
 DA:3,1
-DA:4,1
+DA:4,2
 DA:5,0
 DA:7,1
 DA:8,1