@rfourquet 2025-04-16
> The cmdlineargs test failures seem likely unrelated

So they actually consistently fail. For example at line 497 of the "test/cmdlineargs.jl" file, there is  `@test occursin(expected, got)` failing with this tuple being equal to 
```
 ("SF:/data/src/julia/test/testhelpers/coverage_file.jl\nDA:3,1\nDA:4,1\nDA:5,0\nDA:7,1\nDA:8,1\nDA:9,3\nDA:10,5\nDA:11,1\nDA:12,1\nDA:14,0\nDA:17,1\nDA:18,1\nDA:19,1\nDA:20,1\nDA:22,1\nLH:13\nLF:15\nend_of_record\n",

 "SF:/data/src/julia/test/testhelpers/coverage_file.jl\nDA:3,1\nDA:4,2\nDA:5,0\nDA:7,1\nDA:8,1\nDA:9,3\nDA:10,5\nDA:11,1\nDA:12,1\nDA:14,0\nDA:17,1\nDA:18,1\nDA:19,1\nDA:20,1\nDA:22,1\nLH:13\nLF:15\nend_of_record\nSF:/data/src/julia/usr/share/julia/stdlib/v1.13/Random/src/Random.jl\nDA:141,1\nDA:255,2\nDA:258,1\nDA:260,2\nLH:4\nLF:4\nend_of_record\nSF:/data/src/julia/usr/share/julia/stdlib/v1.13/Random/src/Xoshiro.jl\nDA:214,1\nDA:215,1\nDA:216,1\nDA:217,1\nDA:218,1\nDA:219,0\nDA:220,0\nDA:222,0\nDA:226,1\nDA:227,1\nDA:264,1\nDA:265,1\nDA:266,1\nDA:267,1\nDA:268,1\nDA:269,1\nDA:270,1\nDA:271,1\nDA:272,1\nDA:273,1\nDA:274,1\nDA:275,0\nLH:18\nLF:22\nend_of_record\nSF:/data/src/julia/usr/share/julia/stdlib/v1.13/Random/src/generation.jl\nDA:221,1\nDA:360,1\nDA:364,1\nDA:365,1\nDA:366,1\nDA:367,0\nDA:368,1\nDA:371,1\nDA:375,1\nDA:376,1\nDA:377,1\nDA:378,2\nDA:381,1\nLH:12\nLF:13\nend_of_record\n")
```

The 2nd "DA:" line do not match: it's "4,1" in `expected`, but `4,2` in `got`. I've no idea what this means, are these coverage numbers? In this instance, can I just regenerate the reference "converage_file.info" ?