@vtjnash 2025-04-16
`DA:4,2` means line 4 of file SF got executed twice (instead of once). It is probably fine to just change the reference file to match since you changed inlining, it might now see the line as being run at different points and counting that as being reached twice