@oscardssmith 2025-04-23
@rfourquet do you need help fixing the test? I don't want this branch to be forgotten as it is a quite nice perf win.